portal/app/requirements.txt
responder==2.0.7
redis==3.5.3
six==1.16.0
Flask==2.0.1
Werkzeug==2.0.1
Authlib==0.15.4
constants==0.6.0
python-dotenv==0.18.0
